Unwind circular reference
