+++ src/handlers/events/delivery/deploy/k8s/RequestDeployOnSuccessStatus.ts
@@ -20,10 +20,11 @@ import { GitHubRepoRef } from "@atomist/automation-client/operations/common/GitH
 import { ProjectOperationCredentials, TokenCredentials } from "@atomist/automation-client/operations/common/ProjectOperationCredentials";
 import { RemoteRepoRef } from "@atomist/automation-client/operations/common/RepoId";
 import {
-    currentPhaseIsStillPending, GitHubStatusAndFriends, Phases, PlannedPhase, previousPhaseSucceeded,
+    currentPhaseIsStillPending, GitHubStatusAndFriends, Phases, PlannedPhase,
 } from "../../../../../common/phases/Phases";
 import { OnAnySuccessStatus } from "../../../../../typings/types";
 import { createStatus } from "../../../../../util/github/ghub";
+import { previousPhaseSucceeded } from "../../../../../common/phases/phaseOrdering";
 
 export type K8Target = "testing" | "production";
